Add a Makefile to allow use of make
This allows us to fail compilation if there are any warnings in the
build, like -WError for gcc. Some warnings against this were spotted
online, so now all that is needed is to remove the setting of this
environment variable, if we run into any "warnings as error" problems
due to compiler upgrades that may deprecate existing functions.