Recommend cargo test instead of make test
cargo rebuilds much slower when the environment variable `RUSTFLAGS=-Dwarnings`
is set. So it is much nicer to run tests using cargo test than to use make test. We could
of course also redefine the Makefile recipe behavior but we keep it as it is now to avoid
warnings-producing code from being committed.