Adjust property-based testing of Lagrange interpolation
If we make the interpolation with 20 points, we run into overflow, so we
better avoid that.

This probably is also true for the benchmarks. So they are also fixed
accordingly. Overflows are not caught in benchmarks, since they run in
"release" mode. So we should be a bit conservative about what we allow
there.