Factor out function to check uniqueness of elements
This function is factored out to a new file: utils.rs for the lack of a
better name. We need to start the process of structuring this code base
(read: mess), and this seems like a good starting point. utils.rs will
probably grow over time, and some of the code that is currently nested
under some directory will probably be moved higher up in the directory
tree structure as it is shared across many files.