Switch prime field class from i64 to i128
We might want to allow any integers to be used here such that
PrimeFieldElement -> PrimeFieldElement<T: Num>. But for now, we just use
128-bit numbers. I should look into whether the compiler generates
efficient code, when type parameters are used. I suspect that it *does*
since the types can be determined compile-time and the functions can
thus be adjusted to the known bit lengths.