Add fast pol eval and interpolate
These functions can evaluate and interpolate polynomials in n*log(n)
time in the size of the values and coefficients. They are just wrappers
around the NTT and inverse NTT but I figured it was nice to have these
wrappers, so you don't need a ph.d. in mathematics to use this codebase.