Improve test of low-degree prover
In a run of the property-based test for a 1023 degree polynomial
the test failed to catch the error when y[3] was set to 100 and the
low-degree test passed when it shouldn't. Maybe by chance this value
*was* actually 100, or maybe the security parameter of 40 was too low.
This commit increases the security parameter to 80 and ensures that y[3]
is set to something different than it was before by flipping its last
bit using xor. This *should* prevent this failed test run from happening
again.