Fix issue with boundary quotient
A bad linear interpolation and a wrong value was used for the boundary
zerofier and the boundary linear interpolation (interpolating the input
and output values). This was in part because a bad value for xlast was
used, since the correct xlast value was shadowed by a wrong value. Part
of this mistake wouldn't have occurred if the linter did not allow
shadowing. We should fix it so that it doesn't. Shadowing sucks.