Factor logic to calculate transition zerofier polynomials
This is factored out because this functionality is needed by both the
prover and the verifier. The function returns a pair of polynomials,
one for the numerator, one for the denominator because this allows
the caller to evaluate (multiply and divide with this pair) in the
order that is most computationally effective.