Add linear equation derivation function to Polynomial namespace
This will allow us to make the STARK more general, as this functionality was
until now used from the PrimeFieldElementBig namespace. This should allow
us (at least down the line) to make a generic STARK, where the codeword
element types are only restricted by trait bounds, not to a specific type.