Add STARK verification function
Moves the logic written in the STARK unit test `mimc_big_test` to
a separate function thus implementing a MIMC STARK verifier. There
are still some open questions when it comes to the `transcript`
byte vector, and serialization/deserialization. We should probably
take inspiration from the low-degree tester for this.

This closes #10.