Make failing FRI test input far from low-degree polynomial
The FRI proof system ensures only that the code word is far from (in
terms of Hamming distance) a low-degree polynomial. This means that a
test of a code word of e.g. length 2^14 must change many of the values
away from a low-degree polynomial, and not just, as previously, change
*one* value away from a low-degree polynomial.

This closes #13 with solution 2, which is to "Fix the test".