Ensure FRI checks unique locations in each round
This ensures that the pseudo-random numbers generated for each round in
FRI are all unique. This requires that the random number generator
returns unique numbers in each probe, and that the FRI iterator, that
halves the code word in each round, stops before the codeword becomes
smaller than the number of colinearity checks. So the last codeword must
not be smaller than the number of colinearity checks. This implies that
the codeword in the last round will be a constant (degree 0) as
previously, but will instead have *some* degree. With this commit, that
logic is now handled in both the prover and the verifier.

This change reduces the number of rounds if the expansion factor is
greater than the number of colinearity checks, and this will always be
the case for the parameters that aszepieniec is leaning towards.

This closes #12.