Separate logic to get number of rounds and max degree of last codeword
Given the size of the input codeword to FRI, the max degree of the
codeword, and the number of colinearity checks, we can calculate the
number of rounds (codeword size halvings) and the maximum degree of the
last codeword. This logic is with this commit separated out into its own
function, as this logic was present to places in the code previously.
Also added a test of this function.