Add method for multiplying polynomial with x^n
This is used when building a linear combination of polynomials that go
into the FRI. In a STARK proof, the FRI is only run on one polynomial.
This polynomial is constructed as a linear combination of polynomials
and of polynomials multiplied with x^n to force them to have the right
degree. This is a useful helper method to multiply a polynomial with
x^m.