Refactor and bump Cargo dependencies
- Limit Cargo dependencies to 'MAJOR.MINOR' (no '.PATCH')
- Continue to use [Caret requirements][caret-reqs]
- Update version bounds as consequence of loosening
- Remove comments from Cargo.toml
- Sort Cargo.toml using [`cargo sort`][cargo-sort]

[caret-reqs]: https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#caret-requirements
[cargo-sort]: https://github.com/DevinR528/cargo-sort

When we limit to 'MAJOR.MINOR', we allow for patch updates. We could
have a stricter policy, but until we do, this change favors consistency.

The comments were removed because when auto-formatting Cargo.toml, the
indentation of inline comments positions them awkwardly. `cargo sort` as
a pre-commit hook was not added.