Remove unneccessary `Rem` trait requirement from polynomial elements
`Polynomial` is mainly used for field elements, so the remainder of any
division will always be zero. Remainder/modulus is not used anywhere in
this file, so we don't need to require it. Not requiring it means that
we don't have to implement a meaningless `Rem` trait for e.g.
XFieldElement, as the `rem` function for this type would always be 0, as
it would for all fields.