First complete Stark.prove() for BFieldElement/XFieldElement
A lot of Stark parameters were moved out of the struct and fed to .prove().

Some choices were made wrt. randomizers that aren't documented yet.

Preprocessing was removed as a consequence.

A lot of assertions were added; reduce/move them out eventually.

A lot of cleaning up and refactoring is warranted; some points:
 - Salted Merkle trees: Use these and assess how big the proofs get.
 - LeaflessAuthenticationPaths: If we use these, we need to inject
   values separately.