Add memoization for polynomial multiplication in symbolic evaluation
These two memoization techniques together give a factor 7.3 speedup for
symbolic evaluation when running the benchmark on the medium-sized
rescue prime input where a STARK proof is generated in the generic STARK
prover that uses symbolic evaluation to find the trace polynomials from
the AIR and the trace interpolants.

The idea of memoization came to me as I was trying to sleep.

Thrilled to see that the last couple of days' work paid off!

From mod_pow memoization:
25.40s -> 21.63s

From mul memoization:
21.63s -> 3.46s