Fix error in intermediate result calculation
This fixes an error where the correct `x_powers` power was not applied
to the resulting intermediate result. This was not caught by the rescue
prime test, probably because the `diff_exponents` with non-zero x-powers
where zero everywhere but in the x-powers.

A test was added that would have caught this error.