Add X field mul optimization when multiplying with B field element
aszepieniec suggested that we added an optimization when multiplying an
X field element with an X field element from the B field. This
suggestion lead to a more than factor 2 speedup for NTTs over the X
field. Amazing stuff!

On my slow, weak XPS laptop, this lead to a speedup from 1.3 seconds to
534ms. Very nice.