Add stub for flexible prime field element implementation
This add most of the implementation for a struct that can handle prime
field element for any prime up to 2^256. With some modifications, it
might be able to handle primes up to 2^512 if that should be desired at
one point. This could be achieved by converting the numbers to BigInt
before performing a multiplication.

This struct is added because we want a class that can implement the
`PrimeFieldElement` trait collection which includes the `Copy` trait,
and this cannot be achieved with `BigInt` elements in the struct, so our
`PrimeFieldElementBig` struct should be completely discarded.