Use rayon for a factor 4 speedup of "symbolic evaluation"
This uses rayon to calculate the map reduce of multiplication/sum on
polynomials in symbolic evaluation. On my slow computer this gave a
factor 4 speedup, so the time fell from 79 seconds to 14.5 seconds.

We can probably use this technique other places to, but for some of
them, we might use a data parallel hash map.