Add `Send` and `Sync` traits to U in MPol functions
This adds the requirement that the coefficients in the multivariate
polynomials are of a type that implement the `Send` and `Sync` traits.
These traits allow the contained data types to be used in parallel
implementations thus allowing for a huge speedup in multiple functions,
chief among them anything to do with symbolic evaluation.