Refactor Polynomial<U: Add + ...> into Polynomial<T: PrimeFieldElement>
The main gotcha here was: `#[derive(Clone)]` will complain that `T` does
not have the trait `Clone`, which it shouldn't, because `T::Elem` does.
It seems that the derive logic is a bit crude, and if we just `impl` the
`Clone` trait "manually" (it's still derived by the IDE), it correctly
clones the coefficients.