Increase scope of property-based test of Merkle tree multi proofs
We are bug hunting a bug that we get when shifting from multiproofs with
values, to leafless proofs on salted Merkle trees. This commit so far
represents a failure to hunt down the bug. But I thought it was
important to get more property-based tests of our Merkle trees since
most of our tests of this struct are small and easily verifiable for a
human.