Add salts to Merkle tree used for boundary quotients in STARK proof
This is a step towards achieving 160 bits security for the STARK engine
for the B field.

This leads to a barely noticeable increase in the proof size. The reason
for the small increase is probably that most of the revealed paths are
from FRI which already takes place in the X field, so it doesn't need
salts, and already takes up a lot of space.

Sizes in bytes:
without salts:
36905
38291

With salts
39003
39779
40258

Co-authored-by: Simon Shine <simon@neptune.cash>