Make input/output size of rescue prime variable
The sets the input/output size to 10/5 for the reference STARK
variables. Some adjustment needs to be done to the STARK, as
@aszepieniec has explained that we don't need boundary conditions for
each register. In the absence of boundary conditions, the boundary
condition is calculated as `bq(x) = (ti(x) - 0) / 1`. But in the current
STARK implementation, an explicit boundary condition is needed for each
register.