WIP: Parameterise MerkleMountainRange's Hasher
The purpose: To make a generic interface to hashing so that the same
MerkleMountainRange implementation works for multiple hash functions.
This gives us a chance to benchmark the usage/implementation of hash
functions compared to BLAKE3.

The SimpleHasher interface has both a Value and a Digest type. The Value
type is a type parameter, and the Digest is hardcoded in the `impl`.

The MerkleMountainRange interface is over-complicated because the `impl`
needs access to the Hasher, but the `struct` only needs the Digests.

The problem I'm facing in this commit is...

```
error[E0284]: type annotations needed: cannot satisfy `<_ as simple_hasher::Hasher<blake3::Hash>>::Digest == blake3::Hash`
   --> src/util_types/merkle_mountain_range.rs:178:23
    |
178 |             let mmr = MerkleMountainRange::from_leaf_hashes(&hashes);
    |                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ cannot satisfy `<_ as simple_hasher::Hasher<blake3::Hash>>::Digest == blake3::Hash`
    |
note: required by a bound in `merkle_mountain_range::MerkleMountainRange::<D, H>::from_leaf_hashes`
   --> src/util_types/merkle_mountain_range.rs:20:18
    |
20  |     H: Hasher<D, Digest = D>,
    |                  ^^^^^^^^^^ required by this bound in `merkle_mountain_range::MerkleMountainRange::<D, H>::from_leaf_hashes`
...
23  |     pub fn from_leaf_hashes(leaf_hashes: &[D]) -> Self {
    |            ---------------- required by a bound in this

For more information about this error, try `rustc --explain E0284`.
```