Make simple_hasher compatible with RescuePrime
The `AsBytes` trait is replaced with `ToDigest`: what `blake3::Hash`
needed an internal byte array for was to seed its `blake3::Hasher`.

In the context of RescuePrime, we input field elements. For the purpose
of MerkleTree and MerkleMountainRange, since we always hash two digests
into one, the trivial implementation of `ToDigest<Value>` in each case
is the identity function.

RescuePrime was recently extended with multi-word input/output, but
because it wasn't used yet, it is fixed to 10-5 where we probably want
it to be flexible. In the meantime, the `Hasher` impl pads the input
and reduces the output.