Change RP enforcement to accept input lengths *up to* max length
Previously RP enforced that the input length match *exactly* the
expected input length for the used set of parameters. With this change,
the input is zero-padded to fit the expected length. The field
`input_length` changes name to `max_input_length` to reflect this
change.

Also adds test that this works since we are dealing with a hash function,
after all -- gotta tread carefully.