Use new hasher trait for MMR
Use @sshine's new hasher trait for MMR and add support for using Rescue
Prime as the hasher function for the MMR implementation.

Adds a test that compares the runtime of MMR with Rescue Prime and with
Blake3. This shows that a Rescue Prime-MMR runs 500 times slower than
that with Blake 3. Personally I thought the difference would be bigger.
This is good news imo.