Add function to get peak heights
This function is added to fix the membership verifier in a later commit.
The problem with the membership verifier is that it must verify against
the *correct* peak, not just any in the peak set, when traversing the
authentication path. I think a function that returns peak heights is a
decent way to achieve this, since peaks will always be uniquely
identified from their height as each peak will have a different height.