Add method to update MMR membership proof when leaf is updated
A membership proof is an authentication path and a leaf index. When a
leaf in the MMR is updated (another leaf than that which the proof is
for), the authentication path may change. It *will* change if the
updated leaf and the leaf for which the membership proof applies
point to the same peak (are part of the same Merkle tree). This new
method handles this mutation of an authentication path.