Add Blake3Hash trait, refactor simple_hasher's trait (hash, hash_pair, hash_many)
The Blake3Hash wrapper is necessary to add the Serialize trait, since it
doesn't have this itself. (This was in progress for a year on GitHub [0]).

We rename `hash_one()` and `hash_two()` because the words "one" and
"two" occur in many places where they refer to field elements; the
naming here is slightly confusing.

[0]: https://github.com/BLAKE3-team/BLAKE3/pull/147