Add a conversion from archival MMR to accumulator MMR
Convenient for a leaf update prover as it will be used to create a leaf
update proof in the archival MMR: The plan is to build an accumulator
MMR from the archival MMR and use the accumulator's prove method for the
proof. This saves a clone call to the archival MMR which can be
gigabytes big.