Add function to batch-verify a list of MMR mutations
Add a method to MMR interface with returns true if the pair (old peaks,
new peaks) is consistent with the appended leafs and the mutated leafs
indicated in the function call.

From #47
The reason for this change is mainly that old peaks and new peaks will
only be known at each block, and not between each MMR manipulation.

The internal structure of verify_batch_update is somewhat complicated
as the membership proofs need to be updated throughout the function
 body. The lowest possible workload is achieved if we first apply
all the leaf mutations and then apply all the leaf appends. It's
OK to do it in this order since no leaf that is appended in a block
can be updated in the same block.