Remove prove_append, prove_update_leaf and `AppendProof`, `UpdateProof`
With this commit, the only possible way to verify a sound MMR
manipulation without mutating the MMR is  is through the
`verify_batch_update` function that is used to verify both appends and
leaf mutations. The reason for this change is that the peer is
unlikely to have intermediate MMR/peaks information, rather the peer has
a "before" state and an "after" state, where the state manipulation
consists of multiple appends and leaf mutations.

Re. #47.