Add warning saying that membership proof must be valid before using `mutate_leaf`
The whitepaper states that we need the new leaf, memership proof, and the
old leaf to update an MMRA with a leaf mutation. But the code only uses
the membership proof and the new leaf, not the old leaf. Which is right?

The code is.

To verify that the membership proof for the old leaf is valid, you **do**
need the old leaf. But to calculate the new MMR and MMRA, you do not.
You only need the new leaf and the membership proof. So the function
signature of `mutate_leaf` stays as it is but we add a warning to this
function stating that it is the caller's responsibility that the
membership proof is valid.

This closes #48.