MMR: Simplify 'HashDigest' type parameter to 'H::Digest'
Because the simple_hasher Hasher trait contains a 'Digest' type, it is
not strictly necessary to parameterise data structures over both.

At the same time, use the 'Blake3Hash' wrapper instead of 'blake3::Hash'
because it allows us to impl custom traits for this type, which was not
otherwise possible because how Rust limits impl'ing for data structures
outside one's crate.