Finish Stark Brainfuck constructor
The biggest non-novelty here is introducing Rc<TableCollection>. Because
PermutationArgument refers to a TableCollection and both of those are
returned, we maintain their relationship with an Rc<> instead of adding
lifetimes to PermutationArgument, since those would propagate out of the
Stark object.

We're not entirely convinced this is the best choice, but it appears to
be the least complex solution.