Add function for creating Merkle trees with no values in leafs
This function is added to give the caller some flexibility about how to
make the hashes that go into the leafs. It means that the caller can
decide how to hash any dataset and then build a Merkle tree from those
hashes. This also means that some functions on the Merkle tree will not
be safe, they will panic, if called on a Merkle tree that does not have
values in its leafs. How do we deal with that?

This function is not (directly) tested yet, but it's now called from the
`from_vec` initializer that is fairly thoroughly tested. It should have
its own test though.