Change some types in the BF STARK engine
Apparently we need a `Rc<RefCell<T>>` type, not just a `Rc<T>` type. The
`Rc<T>` type would not allow the caller in STARK `prove` to get a
mutable reference. Adding `RefCell` fixes this. I only discovered this
problem when I acually ran the prover, so the stub for a a test function
that does this has been added.