Move logic from BF STARK prover method to `table_collection`
Also make `Rc<TableCollection>` into `Rc<RefCell<TableCollection>>`
since the `Rc` was not sufficient and failed whenever I tried to get a
mutable borrow in the STARK prover function. Adding a test that actually
called `prove`, even though this is not completed yet, revealed this
error. I asked on IRC about this problem and I was recommended to use
`Rc<RefCell<T>>` which is also what we are using for the tree structure
in mpolynomial.