Make TableCollection iterable as &dyn TableTrait
Using this particular syntax requires upgrading `edition = "2021"` in
Cargo.toml. There is a way around that, but since the argument to stick
to 2018 was that 2021 wasn't stable, and since it is stable now, we'll
just upgrade.

This `.into_iter()` can be used in a few more places than currently, but
requires more consistent use of `TableTrait`. There is no rush, though.

Remove `as_mpolynomial_vars()` as this function was accidentally slipped
in without being used.