Fix error in Brainfuck VM
The register `is_zero` is 1 iff the register `memory_value` is 0. And 0
otherwise. Since the `memory_value` register is initialized to 0, the
`is_zero` register must be initialized to 1. Prior to this commit it was
initialized to 0. The error was found by running the stark prove/verify
test with the environment variable `DEBUG` set. When this is set, a
check on the boundary quotient degrees is, among other things,
performed. That was how the error was caught.