Add successful attack on current STARK BF Engine
Due to a lacking check on the `is_zero` register, we can always enter a
`[` loop even though we shouldn't. This commit shows how an adversarial
simulator can create a proof that passes the verifier even though it
shouldn't. A fix to this problem will be a fix of the currently-failing
test.

Co-authored-by: Jan Ferdinand Sauer <ferdinand@neptune.cash>