Replace Program with source code in Brainfuck STARK constructor
When the source code and not the compiled program is given as input to
the brainfuck constructor, we avoid the problem where the verifier could
accept a validly executed program in "GOTO Brainfuck"/"Brainfuck
assembler" that did not represent a valid Brainfuck program because the
jump destinations were wrong.

This closes #66.