Make the Merkle Tree function `get_leafless_multi_proof` public
I intend to use it in the rewriting/parallellization of XFRI.

This parallellization of digest calculations in XFRI will require that
we change from using the `from_vec` constructor to build a Merkle tree to
using the `from_digests` constructor for Merkle trees. Using this
`from_digests` method instead will necessitate that we get
authentication paths from the Merkle tree and that we do *not* also
store or attempt to get the values from the Merkle tree. Ergo: We need a
way to get a compressed multi-authentication path from a Merkle tree.