Parallellize first round of hashing in Merkle trees in XFRI
The first round of hasing which calculates the digests in the leaves of
the Merkle trees is parallellized with this commit. In the Brainfuck
STARK prover which we are currently optimizing this lead to a reduction
in time spent in the FRI prover from ~17 seconds to ~13 seconds on my P1
Lenovo with 12 cores. It's possible, though, that the hashing of short
arrays should not be parallellized since it introduces some overhead.
This should probably be investigated.