Fix soundness error in FRI verification
This soundness error was one introduced in the Rust porting of the
Brainfuck-STARK engine, not one in the original Python code.

The soundness error is fixed by linking the linking the Merkle root of
the combination codeword with the codeword being verified in FRI. If
these two do not match, an error is returned. Prior to this commit, the
connection between the FRI codeword and the combination codeword was not
enforced.