Optimize FRI `verify()` and `prove()` (#86)
The new algorithm for `verify()` computes "C" values from "A" and "B"
values.  The "C" values are then reused as next rounds "A" values and
thus we only need to read "B" values from the stream.  This may lower
the payload on proof stream roughly by a factor of 3.  This obsoletes
the co-linearity check of each round as we calculate the "C" to be
exactly co-linear with "A" and "B".  Thus we postpone the check to the
last round, where we check that the final "C" values agree with the
`last_codeword` that we know from the proof stream.

Benchmarks as well as proof stream and memory measurements are yet to be
performed.

- Factor out computation of a co-linear point
- Simplify `verify()`
- Test `get_point_on_line`
- Test FRI for 2x^3 + 1

Co-authored-by: Einar Rasmussen <einar@neptune.cash>
Signed-off-by: Einar Rasmussen <einar@neptune.cash>