Introduce weird optimization in B field `add`
Adding this check in the form of an if-branch results in a significant
speedup for some calculations, specifically Rescue Prime. The NTT also
seems to be sped up with 5-10 %. Apparently it results in a faster add.
Perhaps the multiplication can be avoided some times?

See #89 for a discussion.

Rescue Prime digest calculations.
With check on add:
[459.14 ms 459.91 ms 460.73 ms]

With check on mul:
time: [875.10 ms 877.33 ms 879.60 ms]

With no checks:
time: [838.19 ms 840.20 ms 842.12 ms]

With checks on add and on mul:
time: [489.98 ms 491.28 ms 492.57 ms]