Speed up XLIX hash_pair
This changed construction calls the XLIX permutation directly. This
speeds up the hashing by about 5 %. The benchmark for `hash_pair`
decreased its time from 449ms to 429ms with this change. This
`hash_pair` speed is important as that's how Merkle trees are built and
this is the bottle neck when constructing STARK proofs.