Add four test vectors for Rescue Prime XLIX
These are not really test vectors, as they are generated with this
implementation, so we're not really testing anything yet. I just add
these test vectors because I want to ensure that my rewriting of Rescue
Prime on another branch does not alter the output of this hash function.

These test vectors were generated on `origin/master` before the merge of
`thv/faster-rp`. So they should verify that the changes to the hash
function made on thv/faster-rp do no not affect the output.