Reuse already calculated hash digests in MT verification
When verifying a collection of compact/compressed/partial Merkle authentication
paths, we first build a "partial tree" that contains all digests that
can be derived from the total collection of digests and leaves. Instead
of throwing these digests away when verifying the individual Merkle
authentication paths, with this commit, these digest values from
"partial tree" are reused when checking each individual Merkle
authentication path. This reduces the amount of hashing when verifying
a collection of partial authentication paths by a factor 1/2.

With this commit, FRI verification time for the `TWO_BY_TWO_THEN_OUTPUT`
STARK benchmark is reduced from 67ms to 38ms on my DELL XPS.