Change 'Instruction' type to always be one word each
This is not the best modelling, but it solves a fundamental problem:

Indexing into a Vec<Instruction> will not preserve the low-level
semantics of 'instruction_pointer += 2' for instructions that take two
words. A better solution is to keep the program as a Vec<BFieldElement>
and to only decode the current instruction as needed.