Fix bug where MMRA peak was not updated if associated auth path was empty
The new function `batch_mutate_leaf_and_update_mps` is intended to
update a list of membership proofs and an MMR without doing more hashing
than necessary. There was a bug in my previous implementation though.

If the authentication path for one of the associated leaf mutations was
empty (meaning that the updated leaf is itself a peak), then the peak in
the MMRA was not updated. This was only caught because I wrote a
thorough test. This proves, once again, that you shouldn't write
*anything* without testing it.