Fix issue of false positives for mutated MMR auth paths
Fix false positives error when calculating new authentication paths. If
the new leaf is the same as the previous leaf, hashes towards the peak
will still be calculated. This commit adds a check to see if the newly
calculated digests are actually different from the digests contained in
the previous auth paths/MMR membership proofs before counting those MMR
membership proofs as mutated.

Also adds a test that this is handled correctly.