Add methods to convert BFE to and from [u8; 8]
We add this method since the client needs to be able to generate
database keys from block heights, and block heights are

For now, I picked little-endianness to represent the bytes in a B field
element. The reason I chose that is that I programmed `U32s`, which I
expect to be translated to TVM, as little endian. And I thought it made
sense to chose little endianness. The endianness in both places can be
changed, either together or separately, should the need arise.