Update original blake3-FRI with a/b/c optimization
The optimization where c values for this round are used as a values in
the next round is added to the old, original FRI implementation that
uses blake3 for hashing. This change will make the brainfuck-STARK
faster and produce smaller proofs.

Some tests of the output of the verifier has been added.