Thwart @yczhangstju's attack on Brainfuck STARK engine
Cf. Issue 12 on Triton VM, https://github.com/TritonVM/triton-vm/issues/12,
there is an attack on the memory consistency that is supposed to be
guaranteed by the memory table. If it is not guaranteed that the clock
cycle is monotonically increasing within a region in the memory table (a
region being defined as a constant memory pointer value), then memory
consitency can be broken by rearranging rows in the memory table. To
thwart this attack we insert rows in the memory table such that the
clock cycle always increases by one.

This gives rise to a new base column in the memory table. This new
column is used to indicate if the row is also present in the processor
table or if it's inserted as an interweaved row prove that the next row
that is also present in the processor table has a higher clock cycle
than the previous one.

This new base column is used to modify the transition constraint on the
extension column of the memory table which by this commits goes from
degree 2 to degree 3. These changes means that for some programs the
max degree is now found on the extension column for the memory table
(since the memory table can be quite a lot longer than the processor
table which previously was usually defining for the max degree). So when
we calculate the max degree of all tables we have to take extension
columns into account which wasn't the case previously.