Fix tests after thwarting Yun Tsong's attack
The processor table had to be padded before deriving the memory table
from it. That's why a test with a malicious simulator didn't work
before.

Thus to derive a memory table, you have to:
program -> simulate -> pad processor table -> derive memory table

The last step of deriving the memory table handles the padding of the
memory table such that it gets a length that is a power of 2.