Change consistency constraint check on `interweaved` column to next row
In the check, we need to constrain the `interweaved` value of the next row,
otherwise we do not enforce this constraint on the last row. The rest of
the transition constraints along with the permutation argument from the
processor table guarantees that interweaved column is zero in the 1st
row.