Update some calculations of the running product and their associated constraints
Also derive memory matrix before padding processor table.

During this development it was discovered that the memory value in the
last row of the processor table (before padding) is not constrained if
the program ends with a `>` or a `<`. Since this value cannot be output,
we think that's acceptable. We don't think this unconstrained value has
any effects on the soundness of the terminal value/permutation check.