Increase twenty-first version to 0.1.5
This is a re-release of version 0.1.3:
 - Simplify Lagrange interpolation function interface
 - Add faster Lagrange interpolation and benchmarks

This is equivalent to "rolling 0.1.4 back" without reverting the commit.
First, 0.1.4 is yanked (retracted) from crates.io. Then 0.1.3 is placed
on the branch release-0.1.5, is bumped to version 0.1.5, and is released
on crates.io from this branch.

Subsequently, master can be bumped to 0.2.0 and be released.

This addresses the concern and follows the advice given in #28.