README: Add release protocol
This set of steps describes how a new release is made.

This protocol ensures that backwards-breaking changes are not released
with a "major-compatible" version number, causing projects that depend
on this library to break unexpectedly.

It also ensures better ability to track what changes were made. This
helps the discovery of regressions and gives a better overview of each
release.