Remove split-debuginfo = '...', since this breaks MacOS builds
https://doc.rust-lang.org/cargo/reference/profiles.html#split-debuginfo

The split-debuginfo setting controls the -C split-debuginfo flag which
controls whether debug information, if generated, is either placed in
the executable itself or adjacent to it.

This option is a string and acceptable values are the same as those the
compiler accepts. The default value for this option is unpacked on macOS
for profiles that have debug information otherwise enabled. Otherwise
the default for this option is documented with rustc and is
platform-specific. Some options are only available on the nightly
channel. The Cargo default may change in the future once more testing
has been performed, and support for DWARF is stabilized.