Move stark-rescue-prime to separate crate in Cargo workspace
The Rescue-Prime STARK engine is a proof-of-concept zk-STARK engine
built with cryptography from the twenty-first library. The tutorial for
this code is found at:

https://aszepieniec.github.io/stark-anatomy/

The Rescue-Prime STARK has lived as a module in twenty-first until it is
now being moved into its own crate under a Cargo workspace. This means
the Rescue-Prime STARK is continuously compiled, tested and benchmarked
against twenty-first, but is not distributed in the twenty-first crate
at crates.io:

https://crates.io/crates/twenty-first

Co-authored-by: Alan Szepieniec <alan@neptune.cash>
Co-authored-by: Einar Rasmussen <einar@neptune.cash>
Co-authored-by: sword-smith <thor@neptune.cash>