derive PartialEq, Eq for RescuePrimeRegular
This struct contains no fields and so is trivially equivalent to itself.

As various data structures are parameterised over a Hasher and require
to be PartialEq, Eq, a simple way to achieve this with 'derive' is to
let all type parameters be PartialEq, Eq as well.

This should not be necessary, but it also has no negative consequence.