Switch to Montgomery representation on B field
This leads to a 20-30 % reduction in time for NTT on my P17 machine.

The internal representation of the B field element is a bit more
complicated now, so some of the tests had to be changed as well as they
shouldn't access or have opinions about the internal representation,
only the results of various calculations.