Change height of MMRs to be of type u32
Convert height from a u128 to a u32.

Height could have been a u8 since that would still allow for MMRs up to
size 2^256 but since we want to program the MMR in Triton VM assembler
(TASM) we choose a datatype that's compatible with that architecture.

This commit also ensures that the `leftmost_ancestor` algorithm runs in
constant time. Previously it ran in log_2(N) time.