Rewrite MMR function to mutate leafs to use Merkle Tree logic
Rewrite this function to use the same node/index traversal arithmetic
and logic as the Merkle tree implementation. This is made possible by
the new function to convert a leaf index to a Merkle tree index.

Also: delete a test that checked if MMR accumulators could be modified
if they are empty. I don't think that's a relevant or useful test.