Deprecate STARK tutorial libraries by moving into stark-shared crate
This repository is a multi-crate repository, of which only the crate
twenty-first is released on crates.io. The remaining crates contain
tutorial code. With this change, code only used by tutorials is moved
into a separate crate shared between the two tutorial STARKs.

The shared code that is being moved:
 - ProofStream
 - Fri
 - FriDomain
 - StarkVerifyError
 - MerkleProofError