Move DIGEST_LENGTH to rescue_prime_digest
Multiple hash function implementations compete, and each one defines the
digest length. Since the digest length is a part of the interface for
several things, it is better to keep it in a central place.

Consider renaming the module rescue_prime_digest in the event that Tip5
replaces Rescue-Prime. For example into neptune_digest.