Change SpongeHasher::absorb_init() to ::init(), add ::absorb_repeatedly()
Sponge initialization on Triton VM is currently limited to the combined
`absorb_init` instruction for reasons related to arithmetization (it is
preferable to express the constraints of initializing and performing the
first absorption).

This led me to believe that the best interface for `SpongeHasher` was
also `absorb_init()` for more one-to-one correspondence of implementing
STARK/FRI `verify()` in Rust and Triton VM assembly, respectively.

However, adapting STARK/FRI to use SpongeHasher, this is inconvenient.

Separately, but related:
- Add SpongeHasher::absorb_repeatedly() for ::hash_varlen().